advent_of_python/pyproject.toml
[tool.poetry]
name = "advent-of-code-2023"
version = "0.1.0"
description = ""
authors = ["William Martin <will.st4@gmail.com>"]
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
pyfunctional = "^1.4.3"
numpy = "^1.26.2"
toolz = "^0.12.0"
fastapi = "^0.115.0"
dependency-injector = "^4.42.0"
starlette = "^0.38.5"
uvicorn = "^0.30.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
